% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source.R
\name{create_desc_source}
\alias{create_desc_source}
\title{Create data.frame for optional instruction and descriptions compatible with shiny.survey.}
\usage{
create_desc_source(
  method = "df",
  gsheet_id = NULL,
  gsheet_name = NULL,
  gsheet_sheetname = "Descriptions"
)
}
\arguments{
\item{method}{Which method to use. One of 'df' (default) and 'gsheet'.}

\item{gsheet_id}{id if you want to append the sheet to existing googlesheet}

\item{gsheet_name}{optional name for your googlesheet. Used only
if \code{method == 'gsheet'}}

\item{gsheet_sheetname}{name for the spreadsheet name}
}
\value{
Depends on specified method:
\itemize{
\item{\code{df}: data.frame object}
\item{\code{gsheet}: ID of the created googlesheet}
}
}
\description{
Create data.frame for optional instruction and descriptions compatible with shiny.survey.
}
\details{
Function to help initialize source data.frame to generate
optional instructions and descriptions for quetzio.
It should contain specific columns, which are created here alongside one
placeholder input to help populating next ones.

To use 'gsheet' method, the 'googlesheets4' package needs
to be installed.

You can create the source 'data.frame' yourself, keeping only the columns that
are universally mandatory and specific to used types, though it is
recommended to keep all columns that are created during initialization - the
column classes should be kept.

When using 'googlesheet' as a source, all columns should be kept alongside
their order to ensure correct import.
}
