% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gsheet.R
\name{create_survey_questions_gsheet}
\alias{create_survey_questions_gsheet}
\title{Create googlesheet source compatible with shiny.survey}
\usage{
create_survey_questions_gsheet(name = NULL)
}
\arguments{
\item{name}{optional name for your googlesheet}
}
\description{
Create googlesheet source compatible with shiny.survey
}
