% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source.R
\name{create_survey_source}
\alias{create_survey_source}
\title{Create data.frame source compatible with shiny.survey.}
\usage{
create_survey_source(method = "df", name = NULL)
}
\arguments{
\item{method}{Which method to use. One of 'df' (default) and 'gsheet'.}

\item{name}{optional name for your googlesheet. Used only
if \code{method == 'gsheet'}}
}
\value{
Depends on specified method:
\itemize{
\item{\code{df}: data.frame object}
\item{\code{gsheet}: ID of the created googlesheet}
}
}
\description{
Create data.frame source compatible with shiny.survey.
}
\details{
Function to help initialize source data.frame for survey generation.
It should contain specific columns, which are created here alongside one
placeholder input to help populating next ones.

Name suffix helps to determing for which types of inputs these columns
are used. Variables without suffix are used in creation of all types.
\itemize{
\item{\code{chr_}: textInput}
\item{\code{num_}: numericInput}
\item{\code{select_}: selectizeInput}
\item{\code{radio_}: radioButtons}
\item{\code{mult_}: selectizeInput and radioButtons}
}

To use 'gsheet' method, the 'googlesheets4' package needs
to be installed.

You can create the source 'data.frame' yourself, keeping only the columns that
are universally mandatory and specific to used inputTypes, though it is
recommended to keep all columns that are created during initialization - the
variable classes should be kept.

When using 'googlesheet' as a source, all columns should be kept alongside
their order to ensure correct import.
}
