% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_package}
\alias{.check_package}
\title{Helper function to check namespace}
\usage{
.check_package(package)
}
\arguments{
\item{package}{Character string declaring which package to check for}
}
\description{
Helper function to check namespace
}
