% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_source_df}
\alias{.check_source_df}
\title{Check source df for correct columns}
\usage{
.check_source_df(source_df)
}
\arguments{
\item{source_df}{source of inputs}
}
\description{
Check source df for correct columns
}
\keyword{internal}
