% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_source_list}
\alias{.check_source_list}
\title{Check source list for correct columns}
\usage{
.check_source_list(source_list)
}
\arguments{
\item{source_list}{source of inputs}
}
\description{
Check source list for correct columns
}
\keyword{internal}
