% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.custom_css_handler}
\alias{.custom_css_handler}
\title{function to handle css for quetzio UI css classes}
\usage{
.custom_css_handler(css = NULL, div_id)
}
\arguments{
\item{css}{list containing two character objects named with names of css classes. By default it
will provide these styles:
\itemize{
\item{invalid_input = "color: red; font-style: italic;"}
\item{mandatory_star = "color: red;"}
\item{quetzio_submit = "color: #fff; background-color: #337ab7; border-color: #2e6da4; width: 200px;"}
\item{quetzio_description = "font-size: 0.7em;"}
}
If any of these aren't provided in the list passed to \code{css}, the default values would be added.}

\item{div_id}{character string indicating the div id to use this stylesheet}
}
\description{
function to handle css for quetzio UI css classes
}
\keyword{internal}
