% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.df_to_list}
\alias{.df_to_list}
\title{Create source list from 'data.frame'}
\usage{
.df_to_list(source_df)
}
\arguments{
\item{source_df}{source \code{data.frame} object}
}
\description{
Create source list from 'data.frame'
}
