% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_descriptions.R
\name{.generate_description}
\alias{.generate_description}
\title{generate instructions and item descriptions for quetzio with the use of
source list and 'insertUI'}
\usage{
.generate_description(self)
}
\arguments{
\item{self}{R6 'self' object}
}
\description{
generate instructions and item descriptions for quetzio with the use of
source list and 'insertUI'
}
\keyword{internal}
