% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_mandatory}
\alias{.get_mandatory}
\title{function to get indices of mandatory fields}
\usage{
.get_mandatory(source_list)
}
\arguments{
\item{source_list}{Source list of inputs}
}
\description{
function to get indices of mandatory fields
}
\keyword{internal}
