% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_numeric}
\alias{.get_numeric}
\title{function to get indices of numeric fields}
\usage{
.get_numeric(source_list)
}
\arguments{
\item{source_list}{Source list of inputs}
}
\description{
function to get indices of numeric fields
}
