% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_type}
\alias{.get_type}
\title{function to get indices of numeric fields}
\usage{
.get_type(source_list, type)
}
\arguments{
\item{source_list}{Source list of inputs}
}
\description{
function to get indices of numeric fields
}
\keyword{internal}
