% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.gsheet_to_list}
\alias{.gsheet_to_list}
\title{Create source list from googlesheet}
\usage{
.gsheet_to_list(ss_id, ss_name)
}
\arguments{
\item{ss_id}{googlesheet id}

\item{ss_name}{spreadsheet name}
}
\description{
Create source list from googlesheet
}
