% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.label_mandatory}
\alias{.label_mandatory}
\title{function used to add red asterisk to mandatory fields}
\usage{
.label_mandatory(label)
}
\arguments{
\item{label}{Character string with item label}
}
\description{
function used to add red asterisk to mandatory fields
}
\keyword{internal}
