% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.link_backend}
\alias{.link_backend}
\title{Function that creates backend for the linked surveys}
\usage{
.link_backend(self, private, uneval)
}
\arguments{
\item{self}{the 'self' component of R6 object}

\item{private}{the 'private' component of R6 object}

\item{uneval}{the unevaluated expression to create reactiveValues with
list of questionnaires}
}
\description{
Function that creates backend for the linked surveys
}
