% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.list_to_df}
\alias{.list_to_df}
\title{Create correct data.frame from list}
\usage{
.list_to_df(list)
}
\arguments{
\item{list}{source to be converted into data.frame}
}
\description{
Create correct data.frame from list
}
\keyword{internal}
