% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.merge_linked_answers_to_df}
\alias{.merge_linked_answers_to_df}
\title{Merge answers from linked questionnaires into data.frame}
\usage{
.merge_linked_answers_to_df(answers_object, quetzio_names)
}
\arguments{
\item{answers_object}{the 'self$answers()' object from 'quetzio_link_server'}

\item{quetzio_names}{the names of the surveys in correct order from the
'self$quetzio_names' object from 'quetzio_link_server'}
}
\description{
Merge answers from linked questionnaires into data.frame
}
