% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.modify_arg}
\alias{.modify_arg}
\title{function to modify the chosen call argument before evaluation}
\usage{
.modify_arg(call, arg_name, arg_value)
}
\arguments{
\item{call}{the function call to modify}

\item{arg_name}{argument name}

\item{arg_value}{argument value to set}
}
\description{
function to modify the chosen call argument before evaluation
}
\keyword{internal}
