% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.modify_quetzio_arg}
\alias{.modify_quetzio_arg}
\title{function to modify the 'quetzio_server' arguments}
\usage{
.modify_quetzio_arg(..., link_id)
}
\arguments{
\item{...}{dotdotdot passed from parent function}

\item{link_id}{character string indicating what the value of the argument should be}
}
\description{
function to modify the 'quetzio_server' arguments
}
\keyword{internal}
