% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.null_def}
\alias{.null_def}
\title{function for default values if null or NA}
\usage{
.null_def(x, default)
}
\arguments{
\item{x}{the value to check}

\item{default}{the default value if null or NA}
}
\description{
function for default values if null or NA
}
