% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_questions.R
\name{.populate_from_default}
\alias{.populate_from_default}
\title{Populate missing values in your inputs with default provided data}
\usage{
.populate_from_default(source_list, default_config)
}
\arguments{
\item{source_list}{list containing source data}

\item{default_config}{list containing default configuration}
}
\description{
Populate missing values in your inputs with default provided data
}
\details{
The default_config object can have only one default configuration
per input type specified.
If there were some values specified in the vanilla configuration, they won't
be overwritten by configuration
}
\keyword{internal}
