% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.quetzio_label_update}
\alias{.quetzio_label_update}
\title{Server module handling label updates}
\usage{
.quetzio_label_update(
  self,
  private,
  trigger,
  source_method,
  source_yaml,
  source_gsheet_id,
  source_gsheet_sheetname,
  source_object
)
}
\arguments{
\item{self}{the public element of 'quetzio_server' or 'quetzio_link_server'}

\item{source_method}{character string specifying in what form the source
config file will be provided. Can be either 'gsheet', 'yaml' or 'raw'.
Necessity of other arguments is dependent on this choice}

\item{source_yaml}{path to the source yaml file}

\item{source_gsheet_id}{id of the source googlesheet file}

\item{source_gsheet_sheetname}{name of the source spreadsheet}

\item{source_object}{object of class \code{list} (similiar in structure to
'yaml' source) or \code{data.frame} (similiar in structure to 'googlesheet'
source) to be the source of questions. You can create a sample data.frame
with \code{create_survey_source()}. Needed when \code{source_method == 'raw'}}

\item{tigger}{reactive triggering the update}
}
\description{
Server module handling label updates
}
\keyword{internal}
