% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.quetzio_value_update}
\alias{.quetzio_value_update}
\title{Server module handling value updates}
\usage{
.quetzio_value_update(self, values)
}
\arguments{
\item{self}{R6 self object}

\item{values}{reactive object that triggers the change and contains
new values}
}
\description{
Server module handling value updates
}
\details{
likertRadioButtons aren't currently supported!
}
\keyword{internal}
