% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source.R
\name{.randomize_source}
\alias{.randomize_source}
\title{Function to randomize order of source_list object}
\usage{
.randomize_source(source_list)
}
\arguments{
\item{source_list}{list containing items to generate}
}
\value{
List containing two objects:
\itemize{
\item 'source_list', with input in randomized order
\item 'order', with item indices
}
}
\description{
Function to randomize order of source_list object
}
\keyword{internal}
