% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.read_all_answers}
\alias{.read_all_answers}
\title{Read the Answer data from Google Sheets}
\usage{
.read_all_answers(output_ss, output_sheet)
}
\arguments{
\item{output_ss}{character vector with output googlesheet ID}

\item{output_sheet}{character vector with output spreadsheet name}
}
\description{
Read the Answer data from Google Sheets
}
