% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sanitize_answers}
\alias{.sanitize_answers}
\title{sanitize answers from questionnaire}
\usage{
.sanitize_answers(answers, name = NULL)
}
\arguments{
\item{answers}{list of answers from one questionnaire}

\item{name}{name of the questionnaire}
}
\description{
sanitize answers from questionnaire
}
