% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.save_new_answers}
\alias{.save_new_answers}
\title{Save the answer user input to Google Sheet}
\usage{
.save_new_answers(user_answers, output_ss, output_sheet)
}
\description{
Save the answer user input to Google Sheet
}
