% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.update_label}
\alias{.update_label}
\title{Update label of the shinyInput}
\usage{
.update_label(self, inputId, label, is_mandatory)
}
\arguments{
\item{self}{the public component of the 'quetzio_server' object}

\item{inputId}{id of the input to be updated}

\item{label}{character string with updated label value}

\item{is_mandatory}{logical indicating if the item is mandatory}
}
\description{
Update label of the shinyInput
}
\details{
New function is used to handle input label update for two reasons. Firstly,
specific \code{update*Input} functions need to be called for specific input types,
which would create a need for more verbose code with multiple strategies.
Secondly - and most important - \code{update*Input} functions don't allow
supplementing labels with 'HTML' tags, as everything passed through them as
label is coerced to character. \code{shinyjs::html()} remedies this problem.
}
