% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.yaml_to_list}
\alias{.yaml_to_list}
\title{Generate source list from yaml}
\usage{
.yaml_to_list(yaml_file)
}
\arguments{
\item{yaml_file}{path to the source yaml file}
}
\description{
Generate source list from yaml
}
\keyword{internal}
