% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{likertRadioButtons}
\alias{likertRadioButtons}
\title{Create input for Likert scale questions based on radioButtons}
\usage{
likertRadioButtons(
  inputId,
  label,
  selected = character(0),
  choiceValues,
  choiceNames,
  placeholder = "Select value",
  width = "500px"
)
}
\arguments{
\item{inputId}{}

\item{label}{Label that will be shown to the user}

\item{selected}{Initially selected value. Defaults to 'character(0)', which
will make selection empty on init}

\item{choiceValues}{Vector with numeric values to use. Minimal length is 3. Too
long scale may cause the widget to look not as intended.}

\item{choiceNames}{Character vector, indicating meaning of each choiceValue
after user selects it. If you don't wish to show meaning for every value,
provide character vector of length 2 with meaning of min and max}

\item{placeholder}{Initial text in showing meaning of selected choiceValue.
Meaningful only if 'length(choiceNames) > 2' and 'selected = character(0)'.
Defaults to 'Select value'}

\item{width}{Width of the input in either relative (\%) or absolute (px) values.}

\item{max-width}{Maximum width of the input. Absolute (px) values are recommended,
and usage with unison of relative 'width'. Shouldn't be less than 'width'}
}
\description{
Create input for Likert scale questions based on radioButtons
}
