% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{numInput}
\alias{numInput}
\title{Modified Shiny numericInput}
\usage{
numInput(
  inputId,
  label,
  value = NA,
  placeholder = NULL,
  min = NA,
  max = NA,
  step = NA,
  width = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{step}{Interval to use when stepping between min and max}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[=validateCssUnit]{validateCssUnit()}}.}
}
\value{
A numeric input control that can be added to a UI definition. It should
be noted, that empty input will generate NA, not NULL, as in other shinyInputs
}
\description{
Create an input control for entry of numeric values. Contrary to the vanilla
numericInput, it allows no initial value and placeholder content.
}
\section{Server value}{

A numeric vector of length 1.
}

\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  numInput("obs", "Observations:", placeholder = "Enter value", min = 1, max = 100),
  verbatimTextOutput("value")
)
server <- function(input, output) {
  output$value <- renderText({ input$obs })
}
shinyApp(ui, server)
}

}
