% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{quetzio_examples}
\alias{quetzio_examples}
\title{Example source configurations for questionnaires}
\format{
A list containing 12 source lists for 4 different kinds of config, described
in more detail below:
\describe{
\item{questions_lists: }{Contains question sources for 6 different questionnaires: \itemize{
\item{simple_quetzio: }{Simple questionnaire used in example for 'quetzio_server'}
\item{link_quetzio_1, link_quetzio_2, link_quetzio_3: }{Questionnaires
used in example for 'quetzio_link_server'}
\item{gender_update: }{Questionnaire used in example for 'update_labels' method}
\item{simple_default: }{Simple questionnaire showing default generation}
}}
\item{description_lists: }{Contains description sources for 4 different questionnaires: \itemize{
\item{simple_quetzio: }{Descriptions for 'simple_quetzio'}
\item{link_quetzio_1, link_quetzio_2, link_quetzio_3: }{Descriptions for
corresponding linked questionnaires}
}}
\item{label_update: }{Contains 1 configuration for label update used in example}
\item{default_config: }{Contains 1 default configuration for 'simple_default'
question list}
}
}
\usage{
quetzio_examples
}
\description{
A list containing exemplary configuration for 'quetzio_server' class objects
}
\keyword{datasets}
