% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_module.R
\name{quetzio_link_UI}
\alias{quetzio_link_UI}
\title{Create UI for linked questionnaires}
\usage{
quetzio_link_UI(link_id)
}
\arguments{
\item{link_id}{Character string holding ID for the link module. Needs to be the
same as one provided in 'quetzio_link_server'}
}
\description{
Create UI for linked questionnaires
}
\seealso{
quetzio_link_server
}
