% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quetzio.R
\name{Quetzio_UI}
\alias{Quetzio_UI}
\title{Create UI of your questionnaire}
\usage{
Quetzio_UI(module_id)
}
\arguments{
\item{module_id}{Character string holding ID for the module. Needs to be the
same as one provided for 'Quetzio'}
}
\description{
Create UI of your questionnaire
}
\seealso{
Quetzio
}
