% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuetzioLink.R
\name{QuetzioLink_UI}
\alias{QuetzioLink_UI}
\title{Create UI for linked questionnaires}
\usage{
QuetzioLink_UI(link_id)
}
\arguments{
\item{link_id}{Character string holding ID for the link module. Needs to be the
same as one provided in 'QuetzioLink'}
}
\description{
Create UI for linked questionnaires
}
\seealso{
QuetzioLink
}
