% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{updateLikertRadioButtons}
\alias{updateLikertRadioButtons}
\title{Change the value of a likertRadioButtons input on the client}
\usage{
updateLikertRadioButtons(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  selected = NULL,
  choiceValues = NULL,
  choiceNames = NULL
)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.
Default is getDefaultReactiveDomain()}

\item{inputId}{The id of the input object}

\item{label}{The label to set for the input object}

\item{selected}{The numeric value to set for the object}

\item{choiceValues}{Numeric vector of alternative values to set for the object}

\item{choiceNames}{Character vector of alternative labels to set for the object}
}
\description{
Change the value of a likertRadioButtons input on the client
}
