% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{use_language}
\alias{use_language}
\title{Use {shiny.quetzio} labels}
\usage{
use_language(lan = "en")
}
\arguments{
\item{lan}{Language to use for labels, supported values are : "en" and "pl".}
}
\value{
A language object
}
\description{
See all labels registered with \code{get_labels()},
then set custom text with \code{set_labels()}.
}
\examples{

use_language(lan = "pl")

}
