% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source.R
\name{create_quetzio_source}
\alias{create_quetzio_source}
\title{Create data.frame source compatible with shiny.survey}
\usage{
create_quetzio_source(
  method = "df",
  gsheet_id = NULL,
  gsheet_name = NULL,
  gsheet_sheetname = "Questions"
)
}
\arguments{
\item{method}{Which method to use. One of 'df' (default) and 'gsheet'.}

\item{gsheet_id}{id if you want to append the sheet to existing googlesheet}

\item{gsheet_name}{optional name for your googlesheet. Used only
if \code{method == 'gsheet'}}

\item{gsheet_sheetname}{name for the spreadsheet name}
}
\value{
Depends on specified method:
\itemize{
\item{\code{df}: data.frame object}
\item{\code{gsheet}: ID of the created googlesheet}
}
}
\description{
Create data.frame source compatible with shiny.survey
}
\details{
Function to help initialize source data.frame for quetzio generation.
It should contain specific columns, which are created here alongside one
placeholder input to help populating next ones.

Name suffix helps to determine for which types of inputs these columns
are used. Variables without suffix are used in creation of all types.
\itemize{
\item{\code{chr_}: textInput}
\item{\code{num_}: numericInput}
\item{\code{select_}: selectizeInput}
\item{\code{radio_}: radioButtons}
\item{\code{mult_}: selectizeInput and radioButtons}
}

To use 'gsheet' method, the 'googlesheets4' package needs
to be installed.

You can create the source 'data.frame' yourself, keeping only the columns that
are universally mandatory and specific to used inputTypes, though it is
recommended to keep all columns that are created during initialization - the
column classes should be kept.

When using 'googlesheet' as a source, all columns should be kept alongside
their order to ensure correct import.
}
