% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gsheet.R
\name{create_survey_source}
\alias{create_survey_source}
\title{Create data.frame source compatible with shiny.survey.}
\usage{
create_survey_source(method = "df", name = NULL)
}
\arguments{
\item{method}{Which method to use. One of 'df' (default) and 'gsheet'.}

\item{name}{optional name for your googlesheet. Used only
if \code{method == 'gsheet'}}
}
\value{
Depends on specified method:
\itemize{
\item{\code{df}: data.frame object}
\item{\code{gsheet}: character with ID of created googlesheet}
}
}
\description{
Create data.frame source compatible with shiny.survey.
}
\details{
Function to help initialize source data.frame for survey generation.
It should contain specific columns, which are created here alongside one
placeholder input to help populate next ones.

Name suffix helps to determing for which types of inputs these columns
are used. Variables without suffix are used in creation of all types.
\itemize{
\item{\code{chr_}: textInput}
\item{\code{num_}: numericInput}
\item{\code{select_}: selectizeInput}
\item{\code{radio_}: radioButtons}
\item{\code{mult_}: selectizeInput and radioButtons}
}

To use 'gsheet' method, the 'googlesheets4' package needs
to be installed
}
