% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.custom_css_handler}
\alias{.custom_css_handler}
\title{function to handle css for quetzio UI for custom classes: '.invalid_input' and '.mandatory_star'}
\usage{
.custom_css_handler(
  css = list(invalid_input = "color: red; font-style: italic;", mandatory_star =
    "color: red;"),
  div_id
)
}
\arguments{
\item{css}{list containing two character objects, named 'invalid_input' and 'mandatory_star'. By default it
will provide these styles:
\itemize{
\item{invalid_input = "outline: red; outline-style: dashed; outline-offset: 10px;"}
\item{mandatory_star = "color: red;"}
}}

\item{div_id}{character string indicating the div id to use this stylesheet}
}
\description{
function to handle css for quetzio UI for custom classes: '.invalid_input' and '.mandatory_star'
}
