% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source.R
\name{.df_to_list}
\alias{.df_to_list}
\title{Create source list from 'data.frame'}
\usage{
.df_to_list(source_df, type = "quetzio_source")
}
\arguments{
\item{source_df}{source \code{data.frame} object}

\item{type}{character indicating which strategy to take}
}
\description{
Create source list from 'data.frame'
}
\keyword{internal}
