% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_questions.R
\name{.generate_ui}
\alias{.generate_ui}
\title{Generate all shinyInputs from yaml source file function}
\usage{
.generate_ui(source_list, div_id, module_id, css)
}
\arguments{
\item{source_list}{list object with inputs parameters}

\item{div_id}{Character string declaring id for the div}

\item{module_id}{character string declaring module id}

\item{css}{Character string containing custom css rules for classes
'mandatory_star' and 'invalid_input'}
}
\description{
Generate all shinyInputs from yaml source file function
}
