% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_questions.R
\name{.question_ui}
\alias{.question_ui}
\title{Generate shinyInput code for question}
\usage{
.question_ui(x, inputId, module_ui_id)
}
\arguments{
\item{x}{List to get the params for UI generation}

\item{inputId}{Character string with inputId if not in the x}

\item{module_ui_id}{Character string getting the id for module UI}
}
\description{
Generate shinyInput code for question
}
\keyword{internal}
