% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.save_new_answers}
\alias{.save_new_answers}
\title{Save the answer user input to Google Sheet}
\usage{
.save_new_answers(user_answers, output_ss, output_sheet)
}
\arguments{
\item{user_answers}{the object with inputs to extract}

\item{output_ss}{character vector with output googlesheet ID}

\item{output_sheet}{character vector with output spreadsheet name}
}
\description{
Save the answer user input to Google Sheet
}
