% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.survey_backend}
\alias{.survey_backend}
\title{Function that creates 'backend' for the generated survey}
\usage{
.survey_backend(self, private)
}
\arguments{
\item{self}{R6 'self' object}

\item{private}{R6 'private' object}
}
\description{
Function that creates 'backend' for the generated survey
}
\keyword{internal}
