% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.survey_backend}
\alias{.survey_backend}
\title{Function that creates 'backend' for the generated survey}
\usage{
.survey_backend(
  id,
  source_list,
  mandatory_items,
  numeric_items,
  output_gsheet,
  output_ss,
  output_sheet,
  button_labels
)
}
\arguments{
\item{id}{The module id}

\item{source_list}{The list used to create the UI}

\item{mandatory_items}{Numeric vector with indices of items that are mandatory}

\item{numeric_items}{Numeric vector with indices of items that are numeric}

\item{output_gsheet}{Logical indicating if the survey results have to be automatically
saved into googlesheet}

\item{output_ss}{Character vector with output googlesheet id (if \code{output_gsheet = T})}

\item{output_sheet}{Character vector with output spreadsheet name (if \code{output_gsheet = T})}
}
\description{
Function that creates 'backend' for the generated survey
}
