% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Quetzio_get_df}
\alias{Quetzio_get_df}
\alias{Quetzio_get_df.Quetzio}
\alias{Quetzio_get_df.QuetzioLink}
\title{Get answers from Quetzio in the form of data.frame}
\usage{
Quetzio_get_df(Quetzio)

\method{Quetzio_get_df}{Quetzio}(Quetzio)

\method{Quetzio_get_df}{QuetzioLink}(Quetzio)
}
\arguments{
\item{Quetzio}{object of class \code{Quetzio} or \code{Quetzio_link}}
}
\value{
data.frame
}
\description{
Get answers from Quetzio in the form of data.frame
}
