% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_module.R
\name{quetzio_link_server}
\alias{quetzio_link_server}
\title{Create R6-based server module to generate and hold the state of your
linked questionnaires}
\description{
Create R6-based server module to generate and hold the state of your
linked questionnaires

Create R6-based server module to generate and hold the state of your
linked questionnaires
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{link_id}}{character with id of the quetzio_link_module}

\item{\code{completion}}{reactiveVal object holding the rate of linked
questionnaires completion}

\item{\code{message}}{reactiveVal object holding the list of linked questionnaires
error messages}

\item{\code{answers}}{reactiveVal object holding the list of answers to the
linked questionnaires}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{quetzio_link_server$new()}}
\item \href{#method-clone}{\code{quetzio_link_server$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization of the 'quetzio_link_server' object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quetzio_link_server$new(
  link_id,
  ...,
  output_gsheet = FALSE,
  output_gsheet_id = NULL,
  output_gsheet_sheetname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{link_id}}{character specifying the module ID of created object. Needs
to be identical to the one specified inside every 'quetzio_server' objects
provided to the object.}

\item{\code{...}}{'quetzio_server' objects that are going to be connected by the
link. They are all needed to be named - with the names providing the path
to the answers of their questions.
Moreover, they need to be in initialized state: \code{quetzio_server$new(...)}}

\item{\code{output_gsheet}}{logical: do you wish to save the answers automatically
to the googlesheet. If TRUE, the 'output_gsheet_id' and 'output_gsheet_sheetname'
arguments need to be specified. Defaults to FALSE}

\item{\code{output_gsheet_id}}{id of the output googlesheet file}

\item{\code{output_gsheet_sheetname}}{name of the output spreadsheet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{quetzio_link_server$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
