% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_module.R
\name{quetzio_UI}
\alias{quetzio_UI}
\title{Create UI of your questionnaire}
\usage{
quetzio_UI(module_id)
}
\arguments{
\item{module_id}{Character string holding ID for the module. Needs to be the
same as one provided for 'quetzio_server'}
}
\description{
Create UI of your questionnaire
}
