% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_module.R
\name{survey_module}
\alias{survey_module}
\title{Create R6-based Shiny modules for the app}
\description{
Create R6-based Shiny modules for the app

Create R6-based Shiny modules for the app
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{survey_module$new()}}
\item \href{#method-ui}{\code{survey_module$ui()}}
\item \href{#method-server}{\code{survey_module$server()}}
\item \href{#method-clone}{\code{survey_module$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializing the shiny.survey_module object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survey_module$new(
  source_method,
  source_yaml,
  source_gsheet_id,
  source_gsheet_sheetname,
  module_id = NULL,
  div_id = "form",
  output_gsheet_id = source_gsheet_id,
  output_gsheet_sheetname,
  custom_css = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_method}}{character string specifying where to get the source
config file. Can be either 'gsheet' or 'yaml'.}

\item{\code{source_yaml}}{path to the source yaml file}

\item{\code{source_gsheet_id}}{id of the source googlesheet file}

\item{\code{source_gsheet_sheetname}}{name of the source spreadsheet}

\item{\code{module_id}}{character string with unique id for the module. If not
specified, it will be automatically generated}

\item{\code{div_id}}{character string with unique id for the created div. If not
specified, it will be set to 'form'}

\item{\code{output_gsheet_id}}{id of the output googlesheet file. If not specified,
the same googlesheet as for 'source' will be used}

\item{\code{output_gsheet_sheetname}}{name of the output spreadsheet}

\item{\code{custom_css}}{custom css for classes 'mandatory star' and 'invalid_input'.
If not specified, default look will be used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the 'shiny.survey_module' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
Method to input the UI of your survey into the shinyApp UI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survey_module$ui()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The whole UI of your survey
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
Method to input the UI of your survey into the shinyApp UI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survey_module$server()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The reactiveValues holding
/itemize{
/item{is_done logical vector indicating if the survey has been sent. NA indicates error}
/item{message if catches error during survey sending, character string with error message. Otherwise NULL }
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survey_module$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
